connected versions
