+++ package.json
@@ -3,9 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
     "@firebase/firestore": "^4.3.2",
     "@firebase/storage": "^0.11.2",
     "@meghoshpritam/react-image-file-resizer": "^1.0.3",
+    "@mui/material": "^5.15.7",
     "@react-google-maps/api": "^2.19.2",
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
