+++ src/index.js
@@ -4,11 +4,13 @@ import './index.css';
 import App from './App';
 import * as serviceWorkerRegistration from './serviceWorkerRegistration';
 import reportWebVitals from './reportWebVitals';
+import Generalobservation from './pages/Generalobservation';
+import Projectobservation from './pages/Projectobservation';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <App />
+    <Projectobservation/>
   </React.StrictMode>
 );
