+++ src/pages/Homepage.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+function Homepage() {
+  return (
+    <div className="app">
+      <Navbar />
+      <Map />
+      <Footer />
+    </div>
+  );
+}
+
+export default Homepage;
